Finalize compiler/vm rewrite

+++ src/limits.zig
@@ -1,5 +1,5 @@
 const std = @import("std");
 
 pub const max_stack = 1_000_000;
-pub const max_constants = std.math.maxInt(u8) + 1;
-pub const max_locals = std.math.maxInt(u8) + 1;
+pub const max_constants = std.math.maxInt(u16) + 1;
+pub const max_locals = std.math.maxInt(u16) + 1;
